+++ source/blender/blenkernel/intern/dynamicpaint.c
@@ -4133,7 +4133,7 @@ static void surface_determineForceTargetPoints(PaintSurfaceData *sData, int inde
 		closest_d[0] = 1.0f - closest_d[1];
 
 		/* and multiply depending on how deeply force intersects surface */
-		temp = fabs(force_intersect);
+		temp = fabsf(force_intersect);
 		CLAMP(temp, 0.0f, 1.0f);
 		mul_v2_fl(closest_d, acosf(temp) / (float)M_PI_2);
 	}
