+++ source/blender/editors/curve/editcurve.c
@@ -3864,7 +3864,7 @@ static void switchdirection_knots(float *base, int tot)
 	fp1 = base;
 	fp2 = tempf = MEM_mallocN(sizeof(float) * a, "switchdirect");
 	while (a--) {
-		fp2[0] = fabs(fp1[1] - fp1[0]);
+		fp2[0] = fabsf(fp1[1] - fp1[0]);
 		fp1++;
 		fp2++;
 	}
